# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LaunchMailer API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /send/email:
    post:
      operationId: sendEmail
      tags: [send]
      summary: Send email campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailInput"
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"
  /send/whatsapp:
    post:
      operationId: sendWhatsApp
      tags: [send]
      summary: Send WhatsApp campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendWhatsAppInput"
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"
  /send/retry/{campaignId}:
    post:
      operationId: retryCampaign
      tags: [send]
      summary: Retry failed sends for a campaign
      security:
        - bearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Retry result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get current settings (without exposing secrets)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update settings / API keys
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SendEmailInput:
      type: object
      required: [campaignId]
      properties:
        campaignId:
          type: string
    SendWhatsAppInput:
      type: object
      required: [campaignId]
      properties:
        campaignId:
          type: string
    SendResult:
      type: object
      required: [total, sent, failed]
      properties:
        total:
          type: integer
        sent:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string

    # ─── SETTINGS ─────────────────────────────────────────────────────────────
    Settings:
      type: object
      properties:
        resendConfigured:
          type: boolean
        senderEmail:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        twilioConfigured:
          type: boolean
        twilioWhatsappNumber:
          type: ["string", "null"]
        claudeConfigured:
          type: boolean
        companyName:
          type: ["string", "null"]
        companyLogoUrl:
          type: ["string", "null"]
    SettingsUpdate:
      type: object
      properties:
        resendApiKey:
          type: string
        senderEmail:
          type: string
        senderName:
          type: string
        twilioSid:
          type: string
        twilioAuthToken:
          type: string
        twilioWhatsappNumber:
          type: string
        claudeApiKey:
          type: string
        companyName:
          type: string
        companyLogoUrl:
          type: string